+++ src/main/java/net/geforcemods/securitycraft/compat/embeddium/EmbeddiumCompat.java
@@ -1,11 +1,12 @@
-package net.geforcemods.securitycraft.compat.sodium;
+package net.geforcemods.securitycraft.compat.embeddium;
 
-import me.jellysquid.mods.sodium.client.render.chunk.map.ChunkStatus;
-import me.jellysquid.mods.sodium.client.render.chunk.map.ChunkTrackerHolder;
+import org.embeddedt.embeddium.impl.render.chunk.map.ChunkStatus;
+import org.embeddedt.embeddium.impl.render.chunk.map.ChunkTrackerHolder;
+
 import net.minecraft.client.multiplayer.ClientLevel;
 
-public class SodiumCompat {
-	private SodiumCompat() {}
+public class EmbeddiumCompat {
+	private EmbeddiumCompat() {}
 
 	public static void onChunkStatusAdded(ClientLevel level, int x, int z) {
 		ChunkTrackerHolder.get(level).onChunkStatusAdded(x, z, ChunkStatus.FLAG_HAS_BLOCK_DATA);